Refactored ExportModal to use hooks
