+++ src/routes/components/export/index.ts
@@ -1,1 +1,1 @@
-export { default as ExportModal } from './exportModal';
+export { default as ExportModal } from './ExportModal';
